# src/main/java/com/example/blog/model/dto/LoginRequest.java
package com.example.blog.model.dto;

import jakarta.validation.constraints.NotBlank;

@Data
public class LoginRequest {

    @NotBlank(message = "Username is required")
    private String username;

    @NotBlank(message = "Password is required")
    private String password;
}