# src/main/java/com/example/blog/model/dto/RefreshTokenRequest.java
package com.example.blog.model.dto;

import jakarta.validation.constraints.NotBlank;

@Data
public class RefreshTokenRequest {

    @NotBlank(message = "Refresh token is required")
    private String refreshToken;
}